Allow using copy/paste in ex widgets as well
